COMMERCE-5326 Refactor move PriceTag from commerce-taglib to commerce-frontend-taglib
