+++ modules/apps/commerce/commerce-frontend-taglib/src/main/resources/META-INF/resources/price/init.jsp
