+++ modules/apps/commerce/commerce-product-type-grouped-web/src/main/resources/META-INF/resources/render/view.jsp
@@ -69,7 +69,7 @@ String productContentAuthToken = AuthTokenUtil.getToken(request, plid, CPPortlet
 					<c:when test="<%= cpSku != null %>">
 						<h4 class="sku"><%= HtmlUtil.escape(cpSku.getSku()) %></h4>
 
-						<div class="price"><liferay-commerce:price CPDefinitionId="<%= cpDefinitionId %>" CPInstanceId="<%= cpSku.getCPInstanceId() %>" /></div>
+						<div class="price"><commerce-ui:price CPDefinitionId="<%= cpDefinitionId %>" CPInstanceId="<%= cpSku.getCPInstanceId() %>" /></div>
 
 						<div class="subscription-info"><commerce-ui:product-subscription-info CPInstanceId="<%= cpSku.getCPInstanceId() %>" /></div>
 
@@ -82,7 +82,7 @@ String productContentAuthToken = AuthTokenUtil.getToken(request, plid, CPPortlet
 					<c:otherwise>
 						<h4 class="sku" data-text-cp-instance-sku=""></h4>
 
-						<div class="price"><liferay-commerce:price CPDefinitionId="<%= cpDefinitionId %>" /></div>
+						<div class="price"><commerce-ui:price CPDefinitionId="<%= cpDefinitionId %>" /></div>
 
 						<div class="subscription-info" data-text-cp-instance-subscription-info="" data-text-cp-instance-subscription-info-show></div>
