+++ modules/apps/commerce/commerce-taglib/src/main/java/com/liferay/commerce/taglib/servlet/taglib/internal/servlet/ServletContextUtil.java
@@ -22,7 +22,6 @@ import com.liferay.commerce.order.CommerceOrderValidatorRegistry;
 import com.liferay.commerce.order.status.CommerceOrderStatusRegistry;
 import com.liferay.commerce.price.CommerceProductPriceCalculation;
 import com.liferay.commerce.price.list.service.CommercePriceListLocalService;
-import com.liferay.commerce.product.service.CommerceChannelLocalService;
 import com.liferay.commerce.product.util.CPDefinitionHelper;
 import com.liferay.commerce.product.util.CPInstanceHelper;
 import com.liferay.commerce.util.CommerceWorkflowedModelHelper;
@@ -45,12 +44,6 @@ import org.osgi.service.component.annotations.Reference;
 	enabled = false, immediate = true, service = ServletContextUtil.class
 )
 public class ServletContextUtil {
-
-	public static final CommerceChannelLocalService
-		getCommerceChannelLocalService() {
-
-		return _servletContextUtil._getCommerceChannelLocalService();
-	}
 
 	public static final CommerceWorkflowedModelHelper getCommerceOrderHelper() {
 		return _servletContextUtil._getCommerceOrderHelper();
@@ -122,13 +115,6 @@ public class ServletContextUtil {
 	@Deactivate
 	protected void deactivate() {
 		_servletContextUtil = null;
-	}
-
-	@Reference(unbind = "-")
-	protected void setCommerceChannelLocalService(
-		CommerceChannelLocalService commerceChannelLocalService) {
-
-		_commerceChannelLocalService = commerceChannelLocalService;
 	}
 
 	@Reference(
@@ -224,10 +210,6 @@ public class ServletContextUtil {
 		_servletContext = servletContext;
 	}
 
-	private CommerceChannelLocalService _getCommerceChannelLocalService() {
-		return _commerceChannelLocalService;
-	}
-
 	private CommerceWorkflowedModelHelper _getCommerceOrderHelper() {
 		return _commerceWorkflowedModelHelper;
 	}
@@ -286,7 +268,6 @@ public class ServletContextUtil {
 
 	private static ServletContextUtil _servletContextUtil;
 
-	private CommerceChannelLocalService _commerceChannelLocalService;
 	private ModelResourcePermission<CommerceOrder>
 		_commerceOrderModelResourcePermission;
 	private CommerceOrderStatusRegistry _commerceOrderStatusRegistry;
